% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.location.R
\name{as.location}
\alias{as.location}
\alias{as.location.character}
\alias{as.location.location}
\alias{print.location}
\title{Convert a path or URL to a location object.}
\usage{
as.location(x, y, ...)

\method{as.location}{character}(x, y, ...)

\method{as.location}{location}(x, y, ...)

\method{print}{location}(x, ...)
}
\arguments{
\item{x}{Input.}

\item{...}{Ignored.}
}
\description{
Convert a path or URL to a location object.
}
\examples{
\donttest{
# Style files
as.location("apa")
as.location("teaching-and-learning-in-nursing")
as.location("regenerative-medicine-research")

# A URL
url <- 'http://zotero.org/styles/american-journal-of-political-science'
as.location(url)
}
}

