% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl-package.R
\docType{package}
\name{csl-package}
\alias{csl}
\alias{csl-package}
\title{R client for CSL styles}
\description{
R client for CSL styles
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

