% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale_load.R
\name{csl_locale_load}
\alias{csl_locale_load}
\title{Load a CSL locale}
\usage{
csl_locale_load(input, ...)
}
\arguments{
\item{input}{URL or local file path}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Load a CSL locale
}
\details{
This function fetches the style XML document, and parses it into
a more reasonble R list that's easy to navigate. If you want the raw XML,
see \code{locale_xml}
}
\examples{
\dontrun{
# Load a locale from the CSL github repo
de_DE <- 'https://github.com/citation-style-language/locales/raw/master/locales-de-DE.xml'
res <- csl_locale_load(de_DE)
## Query style information
res$info
res$info$translators
res$info$license
res$info$date_updated
res$dates
res$style_options
res$terms

# Load from a local style file
## just specify the style and we read from the local style files
csl_locale_load(input="fr-FR")
}
}
