% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locales.R
\name{csl_locales}
\alias{csl_locales}
\alias{csl_locale_exists}
\title{List locally stored locales}
\usage{
csl_locales(locale = NULL)

csl_locale_exists(locale)
}
\arguments{
\item{locale}{(character) one locale name}
}
\value{
If \code{locale=NULL}, a list of locales. If \code{locale} is
not \code{NULL}, then a full path to the locale file is returned if the
locale exists.
}
\description{
List locally stored locales
}
\examples{
# setup
csl_cache$cache_path_set("seasl", type = "tempdir")
csl_cache$cache_path_get()

# List all locale files
csl_locales()

# cleanup
csl_cache$delete_all()

\dontrun{
# fetch data first
csl_fetch_locales()

# List all locale files
csl_locales()

# list files
csl_locales("et")
csl_locales("fr-FR")

csl_locale_exists("et")
csl_locale_exists("cc")
csl_locale_exists("fr-FR")
}
}
