% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_style_find.R
\name{csl_style_find}
\alias{csl_style_find}
\title{Search for a CSL style}
\usage{
csl_style_find(x)
}
\arguments{
\item{x}{(character) a full or partial journal name}
}
\value{
if no matches \code{NULL}. otherwise, one or more file paths
to the style file on your machine
}
\description{
Search for a CSL style
}
\examples{
# setup
csl_cache$cache_path_set("seasl", type = "tempdir")
csl_cache$mkdir()
dir.create(file.path(csl_cache$cache_path_get(), "styles"))
an <- system.file('inst/examples/acta-naturae.csl', package = 'seasl')
file.copy(an, file.path(csl_cache$cache_path_get(), "styles/acta-naturae.csl"))

# find a style
csl_style_find(x = "Naturae")

# cleanup
csl_cache$delete_all()

\dontrun{
# fetch styles
csl_fetch_styles()

# single match
csl_style_find(x = "American Journal of Epidemiology")

# many matches
csl_style_find(x = "American Journal")
csl_style_find(x = "pediatrics")
csl_style_find(x = "analysis and prevention")

# no matches
csl_style_find(x = "foo bar")
}
}
