% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_load.R
\name{csl_style_load}
\alias{csl_style_load}
\title{Load a CSL style}
\usage{
csl_style_load(input, ...)
}
\arguments{
\item{input}{URL or local file path}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
named list, including slots for
\itemize{
\item info: basic top level information
\item locale: locale information, if it exists
\item macros: macros, 1 to many
\item citation: the citation format, very messy now as the format is messy,
parsed with \link[xml2:as_list]{xml2::as_list}
\item bibliography: very messy for now, we just run \link[xml2:as_list]{xml2::as_list} on this
element as it's complicated to parse
}
}
\description{
Load a CSL style
}
\details{
This function fetches the style XML document, and parses it into
a more reasonble R list that's easy to navigate. If you want the raw XML,
see \code{\link[=csl_style_xml]{csl_style_xml()}}
}
\examples{
# Load a style from the Zotero style repository
x <- 'http://www.zotero.org/styles/american-journal-of-political-science'
if (crul::ok(x)) {
jps <- csl_style_load(x)

## Query style information
jps$info
jps$locale
jps$macros
jps$citation
jps$bibliography
}

\dontrun{
# fetch styles
csl_fetch_styles()
# Load from a local style file
## just specify the style and we read from the local style files
csl_style_load(input="apa")
csl_style_load("computer-und-recht")
csl_style_load("bulletin-de-correspondance-hellenique")
}
}
