% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_xml.R
\name{csl_style_xml}
\alias{csl_style_xml}
\title{Get XML for a CSL style}
\usage{
csl_style_xml(input, raw = FALSE, ...)
}
\arguments{
\item{input}{(character) URL or local file path. Required.}

\item{raw}{(logical) If \code{FALSE} (default) return parsed XML to class
\code{xml_document}. If \code{TRUE}, get character string of XML.}

\item{...}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
an object of class \code{xml_document}, see \pkg{xml2}
to parse the object
}
\description{
Get XML for a CSL style
}
\details{
This function fetches the style XML document. If you want parsed data,
see \code{\link[=csl_style_load]{csl_style_load()}}.
}
\examples{
\dontrun{
csl_style_xml('http://zotero.org/styles/american-journal-of-political-science')
csl_style_xml('http://zotero.org/styles/american-journal-of-political-science', TRUE)
}
}
