% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{csl_styles}
\alias{csl_styles}
\alias{csl_style_exists}
\title{List locally stored styles}
\usage{
csl_styles(style = NULL)

csl_style_exists(style)
}
\arguments{
\item{style}{(character) Style name}
}
\value{
If \code{style=NULL}, a list of length two, independent and dependent
styles. If \code{style} is not \code{NULL}, then a full path to the style file is
returned if the style exists.
}
\description{
List locally stored styles
}
\examples{
\dontrun{
csl_styles()
csl_styles("apa")
csl_styles("zdm")

csl_style_exists("apa")
csl_style_exists("apaggg")
}
}
