% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getstyleslocales.R
\name{get_styles}
\alias{get_locales}
\alias{get_styles}
\title{Get CSL styles or locales}
\usage{
get_styles(path = NULL, ...)

get_locales(path = NULL, ...)
}
\arguments{
\item{path}{(character) Base path to store files in. Defaults to your home path using
\code{Sys.getenv("HOME")}. Styles go to home/styles/ and locales to home/locales/.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}.}
}
\description{
Get CSL styles or locales
}
\examples{
\dontrun{
# Get styles
get_styles()

# Get locales
get_locales()
}
}

