% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_style.R
\name{load_style}
\alias{load_style}
\title{Load a CSL style}
\usage{
load_style(input)
}
\arguments{
\item{input}{URL or local file path}
}
\description{
Load a CSL style
}
\examples{
\dontrun{
# Load a style from the Zotero style repository
jps <- load_style(input = 'http://zotero.org/styles/american-journal-of-political-science')
## Query style information
jps$info
jps$title
jps$citation_format
jps$links_template
jps$editor
jps$author
}
}

