% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locales.R
\name{locales}
\alias{locale_exists}
\alias{locales}
\title{List locales, locally, or from the repository of locales}
\usage{
locales(locale = NULL)

locale_exists(locale)
}
\arguments{
\item{locale}{(character) Locale name}
}
\value{
If \code{locale=NULL}, a list of locales. If \code{locale} is not NULL, then a
path to the locale file is returned if the locale exists.
}
\description{
List locales, locally, or from the repository of locales
}
\examples{
\dontrun{
locales()
locales("et")
locales("fr-FR")

locale_exists("et")
locale_exists("cc")
locale_exists("fr-FR")
}
}

