% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_load.R
\name{style_load}
\alias{style_load}
\title{Load a CSL style}
\usage{
style_load(input, ...)
}
\arguments{
\item{input}{URL or local file path}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}.}
}
\description{
Load a CSL style
}
\details{
This function fetches the style XML document, and parses it into
a more reasonble R list that's easy to navigate. If you want the raw XML,
see \code{\link{style_xml}}.
}
\examples{
\dontrun{
# Load a style from the Zotero style repository
jps <- style_load(input = 'http://www.zotero.org/styles/american-journal-of-political-science')
## Query style information
jps$info
jps$title
jps$citation_format
jps$links_template
jps$editor
jps$author

# Load from a local style file
## just specify the style and we read from the local style files
style_load(input="apa")
style_load(input="zdm")
}
}

