% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/style_xml.R
\name{style_xml}
\alias{style_xml}
\title{Get XML for a CSL style}
\usage{
style_xml(input, raw = FALSE, ...)
}
\arguments{
\item{input}{URL or local file path}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}.}
}
\description{
Get XML for a CSL style
}
\details{
This function fetches the style XML document. If you want parsed data,
see \code{\link{style_load}}.
}
\examples{
\dontrun{
style_xml('http://zotero.org/styles/american-journal-of-political-science')
style_xml('http://zotero.org/styles/american-journal-of-political-science', TRUE)
}
}

