% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{styles}
\alias{style_exists}
\alias{styles}
\title{List styles, locally, or from the CSL repository of styles}
\usage{
styles(style = NULL)

style_exists(style)
}
\arguments{
\item{style}{(character) Style name}
}
\value{
If \code{style=NULL}, a list of length two, independent and dependent styles.
If \code{style} is not NULL, then a path to the style file is returned if the style
exists.
}
\description{
List styles, locally, or from the CSL repository of styles
}
\examples{
\dontrun{
styles()
styles("apa")
styles("zdm")

style_exists("apa")
style_exists("apaggg")
}
}

