% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.location.R
\name{as.location}
\alias{as.location}
\title{Convert a path or URL to a location object.}
\usage{
as.location(x, type = "style")
}
\arguments{
\item{x}{Input}

\item{type}{(character) One of style (default) or locale}
}
\description{
Convert a path or URL to a location object.
}
\examples{
\donttest{
# Style files
as.location("apa")
as.location("teaching-and-learning-in-nursing")
as.location("regenerative-medicine-research")

# A URL
url <- 'http://zotero.org/styles/american-journal-of-political-science'
as.location(url)
}
}

