% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locales.R
\name{csl_locales}
\alias{csl_locales}
\alias{csl_locale_exists}
\title{List locally stored locales}
\usage{
csl_locales(locale = NULL)

csl_locale_exists(locale)
}
\arguments{
\item{locale}{(character) Locale name}
}
\value{
If \code{locale=NULL}, a list of locales. If \code{locale} is
not \code{NULL}, then a full path to the locale file is returned if the
locale exists.
}
\description{
List locally stored locales
}
\examples{
\dontrun{
csl_locales()
csl_locales("et")
csl_locales("fr-FR")

csl_locale_exists("et")
csl_locale_exists("cc")
csl_locale_exists("fr-FR")
}
}
