% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csl_style_find.R
\name{csl_style_find}
\alias{csl_style_find}
\title{Search for a CSL style}
\usage{
csl_style_find(x)
}
\arguments{
\item{x}{(character) a full or partial journal name}
}
\value{
if no matches \code{NULL}. otherwise, one or more file paths
to the style file on your machine
}
\description{
Search for a CSL style
}
\examples{
\dontrun{
# single match
csl_style_find(x = "American Journal of Epidemiology")

# many matches
csl_style_find(x = "American Journal")
csl_style_find(x = "pediatrics")
csl_style_find(x = "analysis and prevention")

# no matches
csl_style_find(x = "foo bar")
}
}
