% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasl-package.R
\docType{package}
\name{seasl-package}
\alias{seasl-package}
\alias{seasl}
\title{R client for CSL styles}
\description{
R client for Citation Style Language (CSL) styles
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
